include patients age, location and last_encounter in listing
