+++ app/models/site.rb
@@ -1,6 +1,7 @@
 class Site < ActiveRecord::Base
   include AutoUUID
   include Resource
+  include WithLocation
 
   belongs_to :institution
   has_one :user, through: :institution
@@ -30,24 +31,6 @@ class Site < ActiveRecord::Base
       where("prefix LIKE concat(?, '%')", institution_or_site.prefix)
     end
   }
-
-  def location(opts={})
-    @location = nil if @location_opts.presence != opts.presence || @location.try(:geo_id) != location_geoid
-    @location_opts = opts
-    @location ||= Location.find(location_geoid, opts)
-  end
-
-  def location=(value)
-    @location = value
-    self.location_geoid = value.try(:id)
-  end
-
-  def self.preload_locations!
-    locations = Location.details(all.map(&:location_geoid).uniq).index_by(&:id)
-    all.to_a.each do |site|
-      site.location = locations[site.location_geoid]
-    end
-  end
 
   def self.filter_by_owner(user, check_conditions)
     if check_conditions
