+++ spec/controllers/patients_controller_spec.rb
@@ -100,8 +100,8 @@ RSpec.describe PatientsController, type: :controller do
       patient1 = institution.patients.make
       patient2 = institution.patients.make
 
-      patient1.encounters.make start_time: DateTime.new(2016, 1, 14, 0, 0, 0)
-      patient2.encounters.make start_time: DateTime.new(2016, 1, 7, 0, 0, 0)
+      patient1.encounters.make start_time: Time.new(2016, 1, 14, 0, 0, 0)
+      patient2.encounters.make start_time: Time.new(2016, 1, 7, 0, 0, 0)
 
       get :index, last_encounter: '1/10/2016'
